+++ src/menu/utils.js
@@ -1,16 +1,3 @@
-let typeQuery = '';
-let typeTimer;
-
-export function matchQuery (buttons, key) {
-	if (!/^[\w| ]+$/i.test(key)) return;
-	if (typeTimer) clearTimeout(typeTimer);
-	typeTimer = setTimeout(() => typeQuery = '', 300);
-	typeQuery += key;
-	const reg = new RegExp(`^${typeQuery}`, 'i');
-	return buttons.find(b => reg.test(b.text));
-}
-
-
 export function addArias (el) {
 	if (!el) return;
 	el.setAttribute('aria-haspopup', 'true');