+++ tests/Menu-utils.spec.js
@@ -1,58 +1,4 @@
 import * as utils from '../src/menu/utils';
-import { waitForTimeout } from './helpers/utils';
-
-
-describe('Menu-utils - utils.matchQuery', () => {
-
-	let buttons;
-
-	beforeEach(async () => {
-		buttons = [
-			{ text: 'Button 1' },
-			{ text: 'Button 2' },
-			{ text: 'Button 3' },
-			{ text: 'Another button' },
-		];
-		await waitForTimeout(301);
-	});
-
-	it('should return undefined if key starts with a non-word character', () => {
-		const result = utils.matchQuery(buttons, '-');
-		expect(result).toBeUndefined();
-	});
-
-	it('should return undefined if key ends with a non-word character', () => {
-		const result = utils.matchQuery(buttons, 'Button%');
-		expect(result).toBeUndefined();
-	});
-
-	it('should return undefined if key contains non-word characters in the middle', () => {
-		const result = utils.matchQuery(buttons, 'Butt#n');
-		expect(result).toBeUndefined();
-	});
-
-	it('should return undefined if no button text starts with the query', () => {
-		const result = utils.matchQuery(buttons, 'xyz');
-		expect(result).toBeUndefined();
-	});
-
-	it('should return the first button whose text starts with the query', () => {
-		const result = utils.matchQuery(buttons, 'Butt');
-		expect(result).toEqual(buttons[0]);
-	});
-
-	it('should return the first button whose text starts with the query, ignoring case', () => {
-		const result = utils.matchQuery(buttons, 'bUtT');
-		expect(result).toEqual(buttons[0]);
-	});
-
-	it('should return the first button whose text starts with the query, even if there are other buttons with similar text', () => {
-		const result = utils.matchQuery(buttons, 'Ano');
-		expect(result).toEqual(buttons[3]);
-	});
-
-});
-
 
 
 describe('Menu-utils - utils.addArias', () => {
@@ -78,7 +24,6 @@ describe('Menu-utils - utils.addArias', () => {
 	});
 
 });
-
 
 
 describe('Menu-utils - removeArias', () => {